refactor(ruby): split profile fetching and conversion
